Move cache to storage module
